# openapi.yaml
openapi: "3.1.0"
info:
  title: NeuralPagedAttention Benchmark API
  version: "2.0.0"
  description: |
    Session-based RL benchmark for KV-cache eviction policy evaluation.

    ## Quickstart
    1. `POST /sessions` — create a session, receive `session_id`
    2. `POST /sessions/{session_id}/reset` — start an episode
    3. `POST /sessions/{session_id}/step` — advance one tick
    4. `GET  /sessions/{session_id}/state` — poll without stepping
    5. `DELETE /sessions/{session_id}` — clean up

    ## Session TTL
    Sessions expire after **inactivity**, not wall-clock age.
    Every /reset, /step, /state, /render, /score call resets the timer.
    Default inactivity TTL: 600 s (10 min). Configurable per session.

    ## Action space (18 discrete actions)
    | ID | Name |
    |----|------|
    | 0  | evict_largest_free |
    | 1  | evict_largest_vip |
    | 2  | evict_oldest_free |
    | 3  | evict_oldest_vip |
    | 4  | swap_largest_free_to_cpu |
    | 5  | swap_largest_vip_to_cpu |
    | 6  | swap_oldest_free_to_cpu |
    | 7  | swap_oldest_vip_to_cpu |
    | 8  | admit_next_free |
    | 9  | admit_next_vip |
    | 10 | reject_next_free |
    | 11 | reject_next_vip |
    | 12 | preempt_shred_free |
    | 13 | preempt_shred_vip |
    | 14 | preempt_swap_free |
    | 15 | preempt_swap_vip |
    | 16 | garbage_collect |
    | 17 | do_nothing |

    ## Observation vector (20 dims, normalized)
    | Idx | Feature |
    |-----|---------|
    | 0   | gpu_utilization_pct |
...

...
servers:
...
paths:
...
  /sessions:
    post:
      tags: [sessions]
      summary: Create a new isolated agent session
      description: |
        Allocates an isolated environment instance. Returns a `session_id` UUID.
        The inactivity timer starts immediately and resets on every /reset,
        /step, /state, /render, or /score call.
      operationId: createSession
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManySessions"

    get:
      tags: [sessions]
      summary: List all active sessions
      operationId: listSessions
      responses:
        "200":
          description: Active session summaries
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionSummary"
                  total:
                    type: integer
  /sessions/{session_id}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      summary: Get session metadata
      operationId: getSession
      responses:
        "200":
          description: Session metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSummary"
        "404":
          $ref: "#/components/responses/SessionNotFound"
    delete:
      tags: [sessions]
      summary: Terminate and clean up a session
      operationId: deleteSession
      responses:
        "200":
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSessionResponse"
        "404":
          $ref: "#/components/responses/SessionNotFound"

  # ── Environment interaction ──────────────────────────────────────────────────
  /sessions/{session_id}/reset:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [environment]
      summary: Reset environment — start a new episode
      description: |
        Resets state, queues, and reward accumulators.
        Returns the initial observation. **Refreshes the inactivity timer.**
      operationId: resetEnv
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetRequest"
      responses:
        "200":
          description: Episode started; initial observation returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetResponse"
        "404":
          $ref: "#/components/responses/SessionNotFound"
  /sessions/{session_id}/step:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [environment]
      summary: Apply an action and advance one timestep
      description: |
        Executes the action, advances tick, spawns traffic, computes reward.
        **Refreshes the inactivity timer.**
        Call /reset before the first step.
      operationId: stepEnv
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepRequest"
      responses:
        "200":
          description: Timestep result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/SessionNotFound"
  /sessions/{session_id}/state:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [environment]
      summary: Poll current observation without stepping
      description: |
        Returns the current observation and episode metadata.
        Safe to call multiple times. **Refreshes the inactivity timer.**
      operationId: getState
      responses:
        "200":
          description: Current environment state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateResponse"
        "404":
          $ref: "#/components/responses/SessionNotFound"
  /sessions/{session_id}/render:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [environment]
      summary: Human-readable text rendering of current state
      description: |
        Returns a text summary suitable for feeding into an LLM agent as context.
        **Refreshes the inactivity timer.**
      operationId: renderState
      responses:
        "200":
          description: Text rendering
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderResponse"
        "404":
          $ref: "#/components/responses/SessionNotFound"

  # ── Benchmark / scoring ──────────────────────────────────────────────────────
  /sessions/{session_id}/score:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [benchmark]
      summary: Current episode score breakdown
      description: Returns live score metrics. **Refreshes the inactivity timer.**
      operationId: getScore
      responses:
        "200":
          description: Score breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResponse"
        "404":
          $ref: "#/components/responses/SessionNotFound"
  /sessions/{session_id}/dashboard:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [benchmark]
      summary: Per-session live dashboard snapshot
      description: |
        Detailed snapshot for dashboard visualisation: state, observation,
        memory layout, queue sizes, TTL countdown.
        **Refreshes the inactivity timer.**
      operationId: sessionDashboard
      responses:
        "200":
          description: Dashboard snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDashboardResponse"
        "404":
          $ref: "#/components/responses/SessionNotFound"
...
  /reset:
    post:
      tags: [legacy]
      summary: "[DEPRECATED] Reset global env"
      deprecated: true
      description: Use `POST /sessions/{session_id}/reset` instead.
      operationId: legacyReset
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                  enum: [easy, medium, hard]
      responses:
        "200":
          description: Reset result (with Deprecation header)
  /step:
    post:
      tags: [legacy]
      summary: "[DEPRECATED] Step global env"
      deprecated: true
      description: Use `POST /sessions/{session_id}/step` instead.
      operationId: legacyStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: integer
                  minimum: 0
                  maximum: 17
      responses:
        "200":
          description: Step result (with Deprecation header)
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Session UUID returned from POST /sessions
  responses:
    SessionNotFound:
      description: Session does not exist or has expired due to inactivity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManySessions:
      description: Server session limit reached
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CreateSessionRequest:
      type: object
      properties:
        agent_id:
          type: string
          nullable: true
          description: Human-readable agent identifier for leaderboard attribution
          example: "ppo-agent-v3"
        task:
          type: string
          enum: [easy, medium, hard]
          default: easy
          description: Initial task difficulty
        inactivity_ttl_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          default: 600
          description: |
            Seconds of silence before auto-expiry. Timer resets on every
            /reset, /step, /state, /render, /score call.
            Minimum: 60 s. Maximum: 86400 s (24 h). Default: 600 s (10 min).
    SessionCreatedResponse:
      type: object
      required: [session_id, created_at, expires_after_inactivity_seconds]
      properties:
        session_id:
          type: string
          format: uuid
        agent_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        expires_after_inactivity_seconds:
          type: integer
          description: Inactivity timeout for this session
        status:
          type: string
          enum: [idle, running]
        message:
          type: string
          description: Human-readable next-step hint
    SessionSummary:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        agent_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: Projected expiry time based on current inactivity
        inactivity_ttl_seconds:
          type: integer
        inactive_for_seconds:
          type: number
          description: Seconds since last interaction
        status:
          type: string
          enum: [idle, running]
        episodes_completed:
          type: integer
        current_task:
          type: string
        current_tick:
          type: integer
    DeleteSessionResponse:
      type: object
      properties:
        session_id:     { type: string }
        terminated:     { type: boolean }
        episodes_completed: { type: integer }
        final_score:    { type: number, nullable: true }

    # ── Reset / Step ───────────────────────────────────────────────────────────
    ResetRequest:
      type: object
      properties:
        task:
          type: string
          enum: [easy, medium, hard]
          nullable: true
          description: Override task difficulty for this episode only
    ResetResponse:
      type: object
      required: [session_id, observation, done]
      properties:
        session_id:  { type: string }
        observation:
          $ref: "#/components/schemas/ObservationDict"
        reward:      { type: number, example: 0.0 }
        done:        { type: boolean, example: false }
        info:
          type: object
          properties:
            task:        { type: string }
            episode_id:  { type: string }
            tick:        { type: integer, example: 0 }
            max_ticks:   { type: integer }
    StepRequest:
      type: object
      required: [action]
      properties:
        action:
          type: integer
          minimum: 0
          maximum: 17
          description: Discrete action ID (0–17)
        action_id:
          type: integer
          minimum: 0
          maximum: 17
          description: "[DEPRECATED] Alias for action. Use action instead."
          deprecated: true
...
    StateResponse:
      type: object
      properties:
        session_id:      { type: string }
        observation:
          $ref: "#/components/schemas/ObservationDict"
        state:           { type: object }
        gpu_util:        { type: number }
        cpu_util:        { type: number }
        free_queue_len:  { type: integer }
        vip_queue_len:   { type: integer }
...